data_interface_service/requirements.txt
pandas
numpy
fastapi
uvicorn
sqlalchemy
python-multipart
psycopg2-binary
asyncpg
pydantic_settings